Refactor external-readme tests to focus on context building
- Removed tests for `generateExternalReadmeContent` from `external-readme.test.js`, as they are now located in a separate manual test file.
- Updated comments to reflect the new structure and clarify the purpose of the existing tests.
- This change streamlines the test suite and improves clarity regarding the scope of each test file.
