+++ tests/lib/utils/external-readme.test.js
@@ -1,13 +1,10 @@
 /**
  * @fileoverview Tests for lib/utils/external-readme.js (context building only, no template).
- * Full template rendering tests are in tests/local/lib/utils/external-readme.test.js.
+ * Template rendering tests (generateExternalReadmeContent) are in tests/manual/external-readme-template.test.js (excluded from CI).
  */
 
 const path = require('path');
-const {
-  buildExternalReadmeContext,
-  generateExternalReadmeContent
-} = require('../../../lib/utils/external-readme');
+const { buildExternalReadmeContext } = require('../../../lib/utils/external-readme');
 
 const projectRoot = path.resolve(__dirname, '..', '..', '..');
 
@@ -104,49 +101,6 @@ describe('external-readme (context only)', () => {
         authType: 'none'
       });
       expect(ctx.secretPaths).toEqual([]);
-    });
-  });
-
-  describe('generateExternalReadmeContent with Secrets section', () => {
-    const fs = require('fs');
-    const templatePath = path.join(projectRoot, 'templates', 'external-system', 'README.md.hbs');
-
-    it('external-system README template exists (required for generateExternalReadmeContent tests)', () => {
-      expect(fs.existsSync(templatePath)).toBe(true);
-    });
-
-    it('generated README contains Secrets section and aifabrix secret set with key (no kv://) when secretPaths present', () => {
-      const content = generateExternalReadmeContent({
-        systemKey: 'hubspot-test',
-        displayName: 'HubSpot Test',
-        authType: 'apikey'
-      });
-      expect(content).toContain('### Secrets');
-      expect(content).toContain('aifabrix secret set');
-      expect(content).toContain('hubspot-test/apiKey');
-      expect(content).toContain('<your value>');
-      expect(content).not.toMatch(/aifabrix secret set kv:\/\//);
-    });
-
-    it('generated README shows aifabrix secret set with systemKey/key and <your value> for oauth2', () => {
-      const content = generateExternalReadmeContent({
-        systemKey: 'hubspot-demo',
-        displayName: 'HubSpot Demo',
-        authType: 'oauth2'
-      });
-      expect(content).toContain('aifabrix secret set hubspot-demo/clientId <your value>');
-      expect(content).toContain('aifabrix secret set hubspot-demo/clientSecret <your value>');
-      expect(content).not.toMatch(/aifabrix secret set kv:\/\//);
-    });
-
-    it('generated README omits Secrets section when authType none', () => {
-      const content = generateExternalReadmeContent({
-        systemKey: 'no-secrets',
-        displayName: 'No Secrets',
-        authType: 'none'
-      });
-      expect(content).not.toMatch(/### Secrets/);
-      expect(content).not.toContain('aifabrix secret set');
     });
   });
 });
